moved get_label_type() to printer_utils
